Testing and Logic Editing in PaymentController
Added a test to the paymentcontroller test suite, which outlines how to test the membership cards. Changed the logic in cardcontroller to correctly verify card swipes. Added a manual entry method for membership cards in payment controller, which has simulation because there is currently no way on the hardware to notify the observer about manual entry.