Customer adds their own bag to the bagging area
In this use case we must determine what will happen when an external bag is added to the electronic scale while the scanner is still enabled. In such a case, we would design the software to allow the customer to continue scanning with the addition of their own bag. Therefore, we will fix the discrepancy of weight which will happen with the verification of an attendant. A method has been created in the BaggingAreaController which will add the bag weight to the expected weight on the electronic scale. This will result in the current weight on the electronic scale and the current expected weight to match which will in turn enable the system again. In a real world scenario, the attendant would visually verify that it is a personal bag and they would fix the problem using this method. Also, we considered the case when the bag is lighter than the electronic scale sensitivity. In this case, the scanner would not disable.